# blockchain/lib/ebsi-core-services/apis/trusted-issuers-registry-v5/api/openapi.yaml
openapi: 3.1.0
info:
  title: Trusted Issuers Registry API v5
  version: 5.0.0
  description: |
    The Trusted Issuers Registry (TIR) is a generic decentralised registry that maintains information about trusted issuers, including public data, accreditations and other relevant details. This information is stored in the smart contract as Attribute Envelopes (similar to Verifiable Credentials), which can be issued by Trusted Issuers or self-issued. Validation of generic Envelopes (like Verifiable Credentials) occurs outside the EBSI Platform. Issuers can define proxies used during credential verification to check if credentials are still valid or have been revoked. The verification flow is detailed in the [Credential Status Framework](https://hub.ebsi.eu/vc-framework/credential-status-framework).

    To register a DID with the TIR, a Legal Entity (LE) must first generate a DID and DID document, and then request a `VerifiableAuthorisationToOnboard` Verifiable Credential (VC) from the Issuer. After presenting the `VerifiableAuthorisationToOnboard` VC, the LE receives an access token, which allows it to register its DID in the DID registry. To become a Truster Issuer, the LE must also request a `VerifiableAccreditationToAttest` VC and register it into the TIR. Upon presenting this VC, the LE receives an access token that permits it to insert its DID into the TIR.

...

    For more information see:
...
  contact:
    name: EBSI Support - Service Desk
...
  license:
    name: EUPL-1.2
...
paths:
  /trusted-issuers-registry/v5/issuers:
    get:
      operationId: get-issuers
      summary: List issuers
      description: Returns a list of trusted issuers.
      tags:
        - Issuers
      parameters:
        - in: query
          name: page[after]
          schema:
            type: string
          description: Cursor that points to the end of the page of data that has been returned.
        - schema:
            type: integer
          in: query
          name: page[size]
          description: Defines the maximum number of objects that may be returned.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  self:
                    type: string
                    description: URI to issuers
                  items:
                    type: array
                    description: list of issuers with their decentralized identifier
                    items:
                      type: object
                      properties:
                        did:
                          type: string
                          description: Issuer's DID
                        href:
                          type: string
                          description: Link to issuer
                  total:
                    type: integer
                    description: Total number of items in a collection
                  pageSize:
                    type: integer
                    description: Number of items to be returned per page
                  links:
                    $ref: "#/components/schemas/links"
              examples:
                Success:
                  value:
                    self: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=182&page[size]=10
                    items:
                      - did: did:ebsi:zppVrNT9bBgMqxrJqVEnvyk
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zppVrNT9bBgMqxrJqVEnvyk
                      - did: did:ebsi:zr2rWDHHrUCdZAW7wsSb5nQ
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zr2rWDHHrUCdZAW7wsSb5nQ
                      - did: did:ebsi:zzQar5zYxUCo1bPypeN3tiF
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zzQar5zYxUCo1bPypeN3tiF
                      - did: did:ebsi:z23usJB14wsPNwTds2esnoJt
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:z23usJB14wsPNwTds2esnoJt
                      - did: did:ebsi:zZviptEhFN1r2u18hhi7Hof
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZviptEhFN1r2u18hhi7Hof
                    total: 1816
                    pageSize: 10
                    links:
                      first: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=1&page[size]=10
                      prev: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=181&page[size]=10
                      next: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=182&page[size]=10
                      last: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=182&page[size]=10
        "400":
          description: Bad Request Error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Bad Paging Request:
                  value:
                    title: Bad Paging Request
                    status: 400
                    detail: PageSize should not be greater than 50
        "500":
          description: Internal Server Error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
  /trusted-issuers-registry/v5/issuers/{did}:
    get:
      operationId: get-issuer
      summary: Get an issuer
      description: Returns a trusted issuer identified by its decentralised identifier (DID).
      tags:
        - Issuers
      parameters:
        - name: did
          in: path
          schema:
            type: string
          examples:
            did:
              description: Valid Pilot DID
              value: did:ebsi:zZeKyEJfUTGwajhNyNX928z
          required: true
        - in: query
          name: version
          description: |
            Version of the endpoint to use. Defaults to the latest version.
          required: false
          schema:
            type: string
            enum:
              - deprecated
              - latest
          example: "deprecated"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                anyOf:
                  - $ref: "#/components/schemas/issuer"
                  - $ref: "#/components/schemas/issuer__deprecated"
              examples:
                Issuer:
                  value:
                    did: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                    attributes: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes
                    hasAttributes: true
                Issuer (deprecated):
                  value:
                    did: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                    attributes:
                      - hash: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
                        body: eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6I1NvMUhJUnQwS3FmMF9CU3dxdjRWTU5na2F4M21DRjlJamNRSndNbkIzYnMifQ.eyJpYXQiOjE2OTE0OTgyNzgsImp0aSI6InVybjp1dWlkOjdmYWIyMmZiLWY1Y2UtNDczNi1hNzQwLWY5YjQzNmMxYmI3NSIsIm5iZiI6MTY5MTQ5ODI3OCwiZXhwIjoxNzIzMDM0Mjc4LCJzdWIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIl0sImlkIjoidXJuOnV1aWQ6N2ZhYjIyZmItZjVjZS00NzM2LWE3NDAtZjliNDM2YzFiYjc1IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlZlcmlmaWFibGVBdHRlc3RhdGlvbiIsIlZlcmlmaWFibGVBY2NyZWRpdGF0aW9uIiwiVmVyaWZpYWJsZUFjY3JlZGl0YXRpb25Ub0F0dGVzdCJdLCJpc3N1ZXIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsImlzc3VhbmNlRGF0ZSI6IjIwMjMtMDgtMDhUMTI6Mzc6NThaIiwiaXNzdWVkIjoiMjAyMy0wOC0wOFQxMjozNzo1OFoiLCJ2YWxpZEZyb20iOiIyMDIzLTA4LTA4VDEyOjM3OjU4WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0wOC0wN1QxMjozNzo1OFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6IiwiYWNjcmVkaXRlZEZvciI6W3sic2NoZW1hSWQiOiJodHRwczovL2FwaS1waWxvdC5lYnNpLmV1L3RydXN0ZWQtc2NoZW1hcy1yZWdpc3RyeS92My9zY2hlbWFzL3ozTWdVRlVrYjcyMnVxNHgzZHY1eUFKbW5ObXpERmVLNVVDOHg4M1FvZUxKTSIsInR5cGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiVmVyaWZpYWJsZUF0dGVzdGF0aW9uIiwiVmVyaWZpYWJsZUF1dGhvcmlzYXRpb25Gb3JUcnVzdENoYWluIl19XSwicmVzZXJ2ZWRBdHRyaWJ1dGVJZCI6ImNmODllNDdlMjYwZWRhNTNiYjE5OTA5NmNhOTczNDc3OTBjODRkMmM4YTVmMmNlM2Y0MTFlYWJhMTRkMDAwM2UifSwiY3JlZGVudGlhbFNjaGVtYSI6eyJpZCI6Imh0dHBzOi8vYXBpLXBpbG90LmVic2kuZXUvdHJ1c3RlZC1zY2hlbWFzLXJlZ2lzdHJ5L3YzL3NjaGVtYXMvempWRk52YkVCUEFyM2E3MjREdHRpb1pwZ1ptTnI3NUJCdFJ6WnFrN3BrRGUiLCJ0eXBlIjoiRnVsbEpzb25TY2hlbWFWYWxpZGF0b3IyMDIxIn19LCJpc3MiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiJ9.AeYv7qp8OE-cVvcSDmDo6X8u8sgCZ40hVE2flaLQSg82wlgS_yGp2FOKjr-nKwflEKoY8LBX-6fQfvFrEpjMnw
                        issuerType: RootTAO
                        tao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                        rootTao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
        "400":
          description: Bad Request
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Invalid {did} Parameter:
                  value:
                    title: Bad Request
                    status: 400
                    detail: The format of {did} parameter is not valid
        "404":
          description: Not found
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DID Not Found:
                  $ref: "#/components/examples/NotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
  /trusted-issuers-registry/v5/issuers/{did}/attributes:
    get:
      operationId: get-issuer-attributes
      summary: List attributes
      description: Returns a list of attributes of a trusted issuer.
      tags:
        - Issuers
      parameters:
        - name: did
          in: path
          schema:
            type: string
          examples:
            did:
              description: Valid Pilot DID
              value: did:ebsi:zZeKyEJfUTGwajhNyNX928z
          required: true
        - in: query
          name: page[after]
          schema:
            type: string
          description: Cursor that points to the end of the page of data that has been returned.
        - schema:
            type: integer
          in: query
          name: page[size]
          description: Defines the maximum number of objects that may be returned.
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  self:
                    type: string
                    description: URI to attributes
                  items:
                    type: array
                    description: List of attributes
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Attribute ID. (hash, sha256)
                        href:
                          type: string
                          description: Link to the resource
                  total:
                    type: integer
                    description: Total number of items in a collection
                  pageSize:
                    type: integer
                    description: Number of items to be returned per page
                  links:
                    $ref: "#/components/schemas/links"
              examples:
                Success:
                  value:
                    self: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes?page[after]=1&page[size]=10
                    items:
                      - id: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
                        href: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes/447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
                    total: 1
                    pageSize: 10
                    links:
                      first: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes?page[after]=1&page[size]=10
                      prev: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes?page[after]=1&page[size]=10
                      next: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes?page[after]=1&page[size]=10
                      last: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers/did:ebsi:zZeKyEJfUTGwajhNyNX928z/attributes?page[after]=1&page[size]=10
        "400":
          description: Bad Request Error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Bad Paging Request:
                  value:
                    title: Bad Paging Request
                    status: 400
                    detail: PageSize should not be greater than 50
        "404":
          description: Not found
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DID Not Found:
                  $ref: "#/components/examples/NotFound"
        "500":
          description: Internal Server Error
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
  /trusted-issuers-registry/v5/issuers/{did}/attributes/{attributeId}:
    get:
      operationId: get-issuer-attribute
      summary: Get an attribute
      description: Returns the latest revision of an issuer attribute.
      tags:
        - Issuers
      parameters:
        - name: did
          in: path
          schema:
            type: string
          examples:
            did:
              description: Valid Pilot DID
              value: did:ebsi:zZeKyEJfUTGwajhNyNX928z
          required: true
        - name: attributeId
          in: path
          schema:
            type: string
          examples:
            attributeId:
              description: Valid Pilot attributeId
              value: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/attributeDetails"
              examples:
                Success:
                  value:
                    did: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                    attribute:
                      hash: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
                      body: eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6I1NvMUhJUnQwS3FmMF9CU3dxdjRWTU5na2F4M21DRjlJamNRSndNbkIzYnMifQ.eyJpYXQiOjE2OTE0OTgyNzgsImp0aSI6InVybjp1dWlkOjdmYWIyMmZiLWY1Y2UtNDczNi1hNzQwLWY5YjQzNmMxYmI3NSIsIm5iZiI6MTY5MTQ5ODI3OCwiZXhwIjoxNzIzMDM0Mjc4LCJzdWIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIl0sImlkIjoidXJuOnV1aWQ6N2ZhYjIyZmItZjVjZS00NzM2LWE3NDAtZjliNDM2YzFiYjc1IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlZlcmlmaWFibGVBdHRlc3RhdGlvbiIsIlZlcmlmaWFibGVBY2NyZWRpdGF0aW9uIiwiVmVyaWZpYWJsZUFjY3JlZGl0YXRpb25Ub0F0dGVzdCJdLCJpc3N1ZXIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsImlzc3VhbmNlRGF0ZSI6IjIwMjMtMDgtMDhUMTI6Mzc6NThaIiwiaXNzdWVkIjoiMjAyMy0wOC0wOFQxMjozNzo1OFoiLCJ2YWxpZEZyb20iOiIyMDIzLTA4LTA4VDEyOjM3OjU4WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0wOC0wN1QxMjozNzo1OFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6IiwiYWNjcmVkaXRlZEZvciI6W3sic2NoZW1hSWQiOiJodHRwczovL2FwaS1waWxvdC5lYnNpLmV1L3RydXN0ZWQtc2NoZW1hcy1yZWdpc3RyeS92My9zY2hlbWFzL3ozTWdVRlVrYjcyMnVxNHgzZHY1eUFKbW5ObXpERmVLNVVDOHg4M1FvZUxKTSIsInR5cGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiVmVyaWZpYWJsZUF0dGVzdGF0aW9uIiwiVmVyaWZpYWJsZUF1dGhvcmlzYXRpb25Gb3JUcnVzdENoYWluIl19XSwicmVzZXJ2ZWRBdHRyaWJ1dGVJZCI6ImNmODllNDdlMjYwZWRhNTNiYjE5OTA5NmNhOTczNDc3OTBjODRkMmM4YTVmMmNlM2Y0MTFlYWJhMTRkMDAwM2UifSwiY3JlZGVudGlhbFNjaGVtYSI6eyJpZCI6Imh0dHBzOi8vYXBpLXBpbG90LmVic2kuZXUvdHJ1c3RlZC1zY2hlbWFzLXJlZ2lzdHJ5L3YzL3NjaGVtYXMvempWRk52YkVCUEFyM2E3MjREdHRpb1pwZ1ptTnI3NUJCdFJ6WnFrN3BrRGUiLCJ0eXBlIjoiRnVsbEpzb25TY2hlbWFWYWxpZGF0b3IyMDIxIn19LCJpc3MiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiJ9.AeYv7qp8OE-cVvcSDmDo6X8u8sgCZ40hVE2flaLQSg82wlgS_yGp2FOKjr-nKwflEKoY8LBX-6fQfvFrEpjMnw
                      issuerType: RootTAO
                      tao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                      rootTao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
        "400":
          description: Bad Request
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Invalid {did} Parameter:
                  value:
                    title: Bad Request
                    status: 400
                    detail: The format of {did} parameter is not valid
        "404":
          description: Not found
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DID Not Found:
                  $ref: "#/components/examples/NotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
...
  /trusted-issuers-registry/v5/issuers/{did}/attributes/{attributeId}/revisions/{revisionId}:
    get:
      operationId: get-issuer-attribute-revision
      summary: Get a revision of an attribute
      description: Returns the revision of an issuer attribute.
      tags:
        - Issuers
      parameters:
        - name: did
          in: path
          schema:
            type: string
          examples:
            did:
              description: Valid Pilot DID
              value: did:ebsi:zZeKyEJfUTGwajhNyNX928z
          required: true
        - name: attributeId
          in: path
          schema:
            type: string
          examples:
            attributeId:
              description: Valid Pilot attributeId
              value: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
          required: true
        - name: revisionId
          in: path
          schema:
            type: string
          examples:
            revisionId:
              description: Valid Pilot revisionId
              value: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/attributeDetails"
              examples:
                Success:
                  value:
                    did: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                    attribute:
                      hash: 447867bafab4c4ad26d4f4650e2390f92a831b6343b7c36ecb7c51fee336fa14
                      body: eyJhbGciOiJFUzI1NiIsInR5cCI6IkpXVCIsImtpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6I1NvMUhJUnQwS3FmMF9CU3dxdjRWTU5na2F4M21DRjlJamNRSndNbkIzYnMifQ.eyJpYXQiOjE2OTE0OTgyNzgsImp0aSI6InVybjp1dWlkOjdmYWIyMmZiLWY1Y2UtNDczNi1hNzQwLWY5YjQzNmMxYmI3NSIsIm5iZiI6MTY5MTQ5ODI3OCwiZXhwIjoxNzIzMDM0Mjc4LCJzdWIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsInZjIjp7IkBjb250ZXh0IjpbImh0dHBzOi8vd3d3LnczLm9yZy8yMDE4L2NyZWRlbnRpYWxzL3YxIl0sImlkIjoidXJuOnV1aWQ6N2ZhYjIyZmItZjVjZS00NzM2LWE3NDAtZjliNDM2YzFiYjc1IiwidHlwZSI6WyJWZXJpZmlhYmxlQ3JlZGVudGlhbCIsIlZlcmlmaWFibGVBdHRlc3RhdGlvbiIsIlZlcmlmaWFibGVBY2NyZWRpdGF0aW9uIiwiVmVyaWZpYWJsZUFjY3JlZGl0YXRpb25Ub0F0dGVzdCJdLCJpc3N1ZXIiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiIsImlzc3VhbmNlRGF0ZSI6IjIwMjMtMDgtMDhUMTI6Mzc6NThaIiwiaXNzdWVkIjoiMjAyMy0wOC0wOFQxMjozNzo1OFoiLCJ2YWxpZEZyb20iOiIyMDIzLTA4LTA4VDEyOjM3OjU4WiIsImV4cGlyYXRpb25EYXRlIjoiMjAyNC0wOC0wN1QxMjozNzo1OFoiLCJjcmVkZW50aWFsU3ViamVjdCI6eyJpZCI6ImRpZDplYnNpOnpaZUt5RUpmVVRHd2FqaE55Tlg5Mjh6IiwiYWNjcmVkaXRlZEZvciI6W3sic2NoZW1hSWQiOiJodHRwczovL2FwaS1waWxvdC5lYnNpLmV1L3RydXN0ZWQtc2NoZW1hcy1yZWdpc3RyeS92My9zY2hlbWFzL3ozTWdVRlVrYjcyMnVxNHgzZHY1eUFKbW5ObXpERmVLNVVDOHg4M1FvZUxKTSIsInR5cGVzIjpbIlZlcmlmaWFibGVDcmVkZW50aWFsIiwiVmVyaWZpYWJsZUF0dGVzdGF0aW9uIiwiVmVyaWZpYWJsZUF1dGhvcmlzYXRpb25Gb3JUcnVzdENoYWluIl19XSwicmVzZXJ2ZWRBdHRyaWJ1dGVJZCI6ImNmODllNDdlMjYwZWRhNTNiYjE5OTA5NmNhOTczNDc3OTBjODRkMmM4YTVmMmNlM2Y0MTFlYWJhMTRkMDAwM2UifSwiY3JlZGVudGlhbFNjaGVtYSI6eyJpZCI6Imh0dHBzOi8vYXBpLXBpbG90LmVic2kuZXUvdHJ1c3RlZC1zY2hlbWFzLXJlZ2lzdHJ5L3YzL3NjaGVtYXMvempWRk52YkVCUEFyM2E3MjREdHRpb1pwZ1ptTnI3NUJCdFJ6WnFrN3BrRGUiLCJ0eXBlIjoiRnVsbEpzb25TY2hlbWFWYWxpZGF0b3IyMDIxIn19LCJpc3MiOiJkaWQ6ZWJzaTp6WmVLeUVKZlVUR3dhamhOeU5YOTI4eiJ9.AeYv7qp8OE-cVvcSDmDo6X8u8sgCZ40hVE2flaLQSg82wlgS_yGp2FOKjr-nKwflEKoY8LBX-6fQfvFrEpjMnw
                      issuerType: RootTAO
                      tao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
                      rootTao: did:ebsi:zZeKyEJfUTGwajhNyNX928z
        "400":
          description: Bad Request
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Invalid {did} Parameter:
                  value:
                    title: Bad Request
                    status: 400
                    detail: The format of {did} parameter is not valid
        "404":
          description: Not found
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DID Not Found:
                  $ref: "#/components/examples/NotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
...
  /trusted-issuers-registry/v5/issuers/{did}/proxies/{proxyId}:
    get:
      operationId: get-issuer-proxy
      summary: Get a proxy
      description: Returns an issuer proxy identified by its proxy ID.
      tags:
        - Issuers
      parameters:
        - name: did
          in: path
          schema:
            type: string
          examples:
            did:
              description: Valid Pilot DID
              value: did:ebsi:zZeKyEJfUTGwajhNyNX928z
          required: true
        - name: proxyId
          in: path
          schema:
            type: string
          examples:
            proxyId:
              description: Valid Pilot proxyId
              value: "0x408db59ec66f14efe30eed366ad57a7f6863fbbb18edd3d3bf15837dc3a494a6"
          required: true
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/proxy"
              examples:
                Success:
                  value:
                    prefix: https://example.com
                    headers:
                      authorization: Bearer 6e9b01cb7b2a269be20aaebc17657d7d
                    testSuffix: /creds/1
        "400":
          description: Bad Request
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Invalid {did} Parameter:
                  value:
                    title: Bad Request
                    status: 400
                    detail: The format of {did} parameter is not valid
        "404":
          description: Not found
          content:
            "application/problem+json":
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                DID Not Found:
                  $ref: "#/components/examples/NotFound"
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
              examples:
                Internal Server Error:
                  $ref: "#/components/examples/InternalError"
...
components:
  schemas:
...
    links:
      title: Links
      description: Links model used for pagination.
      type: object
      properties:
        last:
          type: string
          description: URI of the last page
        prev:
          type: string
          description: URI of the previous page
        next:
          type: string
          description: URI of the next page
        first:
          type: string
          description: URI of the first page
      x-examples:
        Links:
          first: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=1&page[size]=2
          prev: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=2&page[size]=2
          next: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=4&page[size]=2
          last: https://api-pilot.ebsi.eu/trusted-issuers-registry/v5/issuers?page[after]=13&page[size]=2
...
    issuer:
      title: Issuer
      type: object
      description: Issuer object.
      properties:
        did:
          type: string
          description: did of the issuer
        attributes:
          type: string
          description: Link to issuer attributes
        hasAttributes:
          type: boolean
          description: This boolean is true when at least one credential is registered. It is false when the issuer is created but none of the credentials has been accepted yet.
    issuer__deprecated:
      title: Issuer (deprecated)
      type: object
      description: Issuer object.
      properties:
        did:
          type: string
          description: did of the issuer
        attributes:
          type: array
          description: Array of attributes
          items:
            $ref: "#/components/schemas/attribute"
...
  examples:
...
    NotFound:
      value:
        title: Not found
        status: 404
        detail: Not found
    InternalError:
      value:
        title: Internal error
        status: 500
        detail: Internal error
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
servers:
...